@@ -6,15 +6,16 @@ group :development, :test do
 
   gem "cancancan"
   gem "pundit"
-  gem "jruby-openssl", platform: :jruby
 
   gem "draper"
   gem "devise"
 
-  gem "rails", "~> 6.1.0"
-  gem "activerecord-jdbcsqlite3-adapter", platform: :jruby
+  gem "rails", "~> 7.1.0"
 
-  gem "webpacker"
+  gem "sprockets-rails"
+  gem "sassc-rails"
+  gem "ransack", ">= 4.1.0"
+  gem "formtastic", ">= 5.0.0"
 end
 
 group :test do
@@ -23,10 +24,10 @@ group :test do
   gem "webrick"
 
   gem "simplecov", require: false # Test coverage generator. Go to /coverage/ after running tests
-  gem "simplecov-cobertura"
+  gem "simplecov-cobertura", require: false
   gem "cucumber-rails", require: false
   gem "cucumber"
-  gem "database_cleaner"
+  gem "database_cleaner-active_record"
   gem "launchy"
   gem "parallel_tests"
   gem "rspec-rails"
@@ -36,6 +37,23 @@ group :test do
   gem "i18n-tasks"
   gem "i18n-spec"
   gem "rails-i18n" # Provides default i18n for many languages
+end
+
+group :release do
+  gem "chandler" # Github releases from changelog
+  gem "octokit"
+end
+
+group :rubocop do
+  gem "rubocop"
+  gem "rubocop-packaging"
+  gem "rubocop-rspec"
+  gem "rubocop-rails"
+end
+
+group :docs do
+  gem "yard" # Documentation generator
+  gem "kramdown" # Markdown implementation (for yard)
 end
 
 gemspec path: "../.."
