@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class Blog::Post < ActiveRecord::Base
+class Blog::Post < ApplicationRecord
   belongs_to :category, foreign_key: :custom_category_id
   belongs_to :author, class_name: "User"
   has_many :taggings
