@@ -1,4 +1,4 @@
 # frozen_string_literal: true
-class Profile < ActiveRecord::Base
+class Profile < ApplicationRecord
   belongs_to :user
 end
