@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class Tag < ActiveRecord::Base
+class Tag < ApplicationRecord
   has_many :taggings
   has_many :posts, through: :taggings
 end
