@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class Tagging < ActiveRecord::Base
+class Tagging < ApplicationRecord
   belongs_to :post, optional: true
   belongs_to :tag, optional: true
 
