@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   class VIP < self
   end
   has_many :posts, foreign_key: "author_id"
