@@ -12,6 +12,7 @@ require_relative "../../tasks/test_application"
 require "#{ActiveAdmin::TestApplication.new.full_app_dir}/config/environment.rb"
 
 require_relative "rails"
+require_relative "../../spec/support/active_support_deprecation"
 
 require "rspec/mocks"
 World(RSpec::Mocks::ExampleMethods)
@@ -57,7 +58,7 @@ DatabaseCleaner.strategy = :truncation
 Cucumber::Rails::Database.javascript_strategy = :truncation
 
 # Warden helpers to speed up login
-# See https://github.com/plataformatec/devise/wiki/How-To:-Test-with-Capybara
+# See https://github.com/heartcombo/devise/wiki/How-To:-Test-with-Capybara
 include Warden::Test::Helpers
 
 After do
@@ -71,7 +72,7 @@ Before do
 end
 
 # Force deprecations to raise an exception.
-ActiveSupport::Deprecation.behavior = :raise
+ActiveAdmin::DeprecationHelper.behavior = :raise
 
 After "@authorization" do |scenario, block|
   # Reset back to the default auth adapter
