@@ -3,7 +3,6 @@ require "active_support/core_ext"
 require "set"
 
 require "ransack"
-require "ransack_ext"
 require "kaminari"
 require "formtastic"
 require "formtastic_i18n"
@@ -27,7 +26,6 @@ module ActiveAdmin
   autoload :ControllerAction, "active_admin/controller_action"
   autoload :CSVBuilder, "active_admin/csv_builder"
   autoload :Dependency, "active_admin/dependency"
-  autoload :Deprecation, "active_admin/deprecation"
   autoload :Devise, "active_admin/devise"
   autoload :DSL, "active_admin/dsl"
   autoload :FormBuilder, "active_admin/form_builder"
@@ -60,6 +58,10 @@ module ActiveAdmin
 
     def application
       @application ||= ::ActiveAdmin::Application.new
+    end
+
+    def deprecator
+      @deprecator ||= ActiveSupport::Deprecation.new("4.0", "active-admin")
     end
 
     # Gets called within the initializer
