@@ -41,7 +41,7 @@ module ActiveAdmin
 
       def predicate_name
         I18n.t(
-          "active_admin.filters.predicates.#{condition.predicate.name}",
+          "ransack.predicates.#{condition.predicate.name}",
           default: ransack_predicate_name)
       end
 
