@@ -21,7 +21,7 @@ module ActiveAdmin
       # Returns the default filter type for a given attribute. If you want
       # to use a custom search method, you have to specify the type yourself.
       def default_input_type(method, options = {})
-        if method =~ /_(eq|equals|cont|contains|start|starts_with|end|ends_with)\z/
+        if method =~ /_(eq|cont|start|end)\z/
           :string
         elsif klass._ransackers.key?(method.to_s)
           klass._ransackers[method.to_s].type
@@ -31,7 +31,7 @@ module ActiveAdmin
           case column.type
           when :date, :datetime
             :date_range
-          when :string, :text
+          when :string, :text, :citext
             :string
           when :integer, :float, :decimal
             :numeric
