@@ -3,8 +3,8 @@ module ActiveAdmin
   module Filters
 
     class Disabled < RuntimeError
-      def initialize
-        super "Can't remove a filter when filters are disabled. Enable filters with 'config.filters = true'"
+      def initialize(action)
+        super "Cannot #{action} a filter when filters are disabled. Enable filters with 'config.filters = true'"
       end
     end
 
@@ -61,7 +61,7 @@ module ActiveAdmin
       #
       # @param [Symbol] attributes The attributes to not filter on
       def remove_filter(*attributes)
-        raise Disabled unless filters_enabled?
+        raise Disabled, "remove" unless filters_enabled?
 
         attributes.each { |attribute| (@filters_to_remove ||= []) << attribute.to_sym }
       end
@@ -73,7 +73,7 @@ module ActiveAdmin
       # @param [Hash] options The set of options that are passed through to
       #                       ransack for the field definition.
       def add_filter(attribute, options = {})
-        raise Disabled unless filters_enabled?
+        raise Disabled, "add" unless filters_enabled?
 
         (@filters ||= {})[attribute.to_sym] = options
       end
