@@ -9,7 +9,7 @@
 #     include Base
 #     include Base::SearchMethodSelect
 #
-#     filter :equals, :greater_than, :less_than
+#     filter :eq, :gt, :lt
 #   end
 #
 module ActiveAdmin
@@ -65,7 +65,7 @@ module ActiveAdmin
 
           def filter_options
             filters.collect do |filter|
-              [I18n.t("active_admin.filters.predicates.#{filter}"), "#{method}_#{filter}"]
+              [I18n.t("ransack.predicates.#{filter}").capitalize, "#{method}_#{filter}"]
             end
           end
 
