@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 module ActiveAdmin
-  VERSION = "2.13.1"
+  VERSION = "3.2.5"
 end
