@@ -102,10 +102,6 @@ module ActiveAdmin
         form_builder.object
       end
 
-      def form_buffers
-        raise "'form_buffers' has been removed from ActiveAdmin::FormBuilder, please read https://github.com/activeadmin/activeadmin/blob/master/docs/5-forms.md for details."
-      end
-
       private
 
       def create_another_checkbox
@@ -131,11 +127,23 @@ module ActiveAdmin
         html_options[:class] ||= "inputs"
         legend = args.shift if args.first.is_a?(::String)
         legend = html_options.delete(:name) if html_options.key?(:name)
-        legend_tag = legend ? "<legend><span>#{legend}</span></legend>" : ""
-        fieldset_attrs = html_options.map { |k, v| %Q{#{k}="#{v}"} }.join(" ")
+        legend_tag = legend ? helpers.tag.legend(helpers.tag.span(legend), class: "fieldset-title") : ""
+        fieldset_attrs = tag_attributes html_options
         @opening_tag = "<fieldset #{fieldset_attrs}>#{legend_tag}<ol>"
         @closing_tag = "</ol></fieldset>"
         super(*(args << html_options), &block)
+      end
+
+      private
+
+      def tag_attributes(html_options)
+        if Rails::VERSION::MAJOR <= 6
+          # Reimplement tag.attributes to backport support for Rails 6.1.
+          # TODO: this can be removed when support for Rails 6.x is dropped
+          helpers.tag.tag_options(html_options.to_h).to_s.strip.html_safe
+        else
+          helpers.tag.attributes html_options
+        end
       end
     end
 
