@@ -15,7 +15,7 @@ RSpec.describe ActiveAdmin::Filters::ActiveFilter do
   let(:post) { Post.create! title: "Hello World", category: category, author: user }
 
   let(:search) do
-    Post.ransack(title_equals: post.title)
+    Post.ransack(title_eq: post.title)
   end
 
   let(:condition) do
@@ -47,7 +47,7 @@ RSpec.describe ActiveAdmin::Filters::ActiveFilter do
   end
 
   it "should pick predicate name translation" do
-    expect(subject.predicate_name).to eq(I18n.t("active_admin.filters.predicates.equals"))
+    expect(subject.predicate_name).to eq(I18n.t("ransack.predicates.eq"))
   end
 
   context "search by belongs_to association" do
@@ -214,7 +214,7 @@ RSpec.describe ActiveAdmin::Filters::ActiveFilter do
     let(:post) { resource_klass.create! title: "Category", author: user }
 
     let(:search) do
-      resource_klass.ransack(title_equals: post.title)
+      resource_klass.ransack(title_eq: post.title)
     end
 
     it "should use the association's primary key to find the associated record" do
