@@ -71,19 +71,19 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
     let(:body) { filter :title }
 
     it "should generate a select option for starts with" do
-      expect(body).to have_selector("option[value=title_starts_with]", text: "Starts with")
+      expect(body).to have_selector("option[value=title_start]", text: "Starts with")
     end
 
     it "should generate a select option for ends with" do
-      expect(body).to have_selector("option[value=title_ends_with]", text: "Ends with")
+      expect(body).to have_selector("option[value=title_end]", text: "Ends with")
     end
 
     it "should generate a select option for contains" do
-      expect(body).to have_selector("option[value=title_contains]", text: "Contains")
+      expect(body).to have_selector("option[value=title_cont]", text: "Contains")
     end
 
     it "should generate a text field for input" do
-      expect(body).to have_selector("input[name='q[title_contains]']")
+      expect(body).to have_selector("input[name='q[title_cont]']")
     end
 
     it "should have a proper label" do
@@ -97,26 +97,26 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
     end
 
     it "should select the option which is currently being filtered" do
-      scope = Post.ransack title_starts_with: "foo"
+      scope = Post.ransack title_start: "foo"
       body = Capybara.string(render_filter scope, title: {})
-      expect(body).to have_selector("option[value=title_starts_with][selected=selected]", text: "Starts with")
+      expect(body).to have_selector("option[value=title_start][selected=selected]", text: "Starts with")
     end
 
     context "with filters options" do
-      let(:body) { filter :title, filters: [:contains, :starts_with] }
+      let(:body) { filter :title, filters: [:cont, :start] }
 
       it "should generate provided options for filter select" do
-        expect(body).to have_selector("option[value=title_contains]", text: "Contains")
-        expect(body).to have_selector("option[value=title_starts_with]", text: "Starts with")
+        expect(body).to have_selector("option[value=title_cont]", text: "Contains")
+        expect(body).to have_selector("option[value=title_start]", text: "Starts with")
       end
 
       it "should not generate a select option for ends with" do
-        expect(body).not_to have_selector("option[value=title_ends_with]")
+        expect(body).not_to have_selector("option[value=title_end]")
       end
     end
 
     context "with predicate" do
-      %w[eq equals cont contains start starts_with end ends_with].each do |predicate|
+      %w[eq cont start end].each do |predicate|
         describe "'#{predicate}'" do
           let(:body) { filter :"title_#{predicate}" }
 
@@ -137,9 +137,9 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
     let(:body) { filter :reason_of_sign_in }
 
     it "should generate a select options" do
-      expect(body).to have_selector("option[value=reason_of_sign_in_starts_with]")
-      expect(body).to have_selector("option[value=reason_of_sign_in_ends_with]")
-      expect(body).to have_selector("option[value=reason_of_sign_in_contains]")
+      expect(body).to have_selector("option[value=reason_of_sign_in_start]")
+      expect(body).to have_selector("option[value=reason_of_sign_in_end]")
+      expect(body).to have_selector("option[value=reason_of_sign_in_cont]")
     end
   end
 
@@ -147,7 +147,7 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
     let(:body) { filter :body }
 
     it "should generate a search field for a text attribute" do
-      expect(body).to have_selector("input[name='q[body_contains]']")
+      expect(body).to have_selector("input[name='q[body_cont]']")
     end
 
     it "should have a proper label" do
@@ -172,6 +172,21 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
           m
         }
         body
+      end
+
+      context "and a statement timeout error occurs" do
+        let(:body) { filter :title, as: :select, collection: ["foo"] }
+        let(:input_super_class) { Formtastic::Inputs::Base::Collections }
+        let(:db_timeout_exception) { ActiveRecord::QueryCanceled.new("ERROR: canceling statement due to statement timeout") }
+        let(:expected_exception_message) { "ERROR: canceling statement due to statement timeout while querying the values for the ActiveAdmin :title filter" }
+
+        before do
+          expect_any_instance_of(input_super_class).to receive(:collection).and_raise(db_timeout_exception)
+        end
+
+        it "should raise a database timeout error with a message indicating which filter was the cause" do
+          expect { body }.to raise_error(ActiveRecord::QueryCanceled, expected_exception_message)
+        end
       end
     end
   end
@@ -221,19 +236,19 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
     let(:body) { filter :created_at }
 
     it "should generate a date greater than" do
-      expect(body).to have_selector("input.datepicker[name='q[created_at_gteq_datetime]']")
+      expect(body).to have_selector("input.datepicker[name='q[created_at_gteq]']")
     end
 
     it "should generate a date less than" do
-      expect(body).to have_selector("input.datepicker[name='q[created_at_lteq_datetime]']")
+      expect(body).to have_selector("input.datepicker[name='q[created_at_lteq]']")
     end
 
     context "with input_html" do
       let(:body) { filter :created_at, input_html: { 'autocomplete': "off" } }
 
       it "should generate provided input html for both ends of date range" do
-        expect(body).to have_selector("input.datepicker[name='q[created_at_gteq_datetime]'][autocomplete=off]")
-        expect(body).to have_selector("input.datepicker[name='q[created_at_lteq_datetime]'][autocomplete=off]")
+        expect(body).to have_selector("input.datepicker[name='q[created_at_gteq]'][autocomplete=off]")
+        expect(body).to have_selector("input.datepicker[name='q[created_at_lteq]'][autocomplete=off]")
       end
     end
 
@@ -241,8 +256,8 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
       let(:body) { filter :created_at, input_html: { 'class': "custom_class" } }
 
       it "should override the default attribute values for both ends of date range" do
-        expect(body).to have_selector("input.custom_class[name='q[created_at_gteq_datetime]']")
-        expect(body).to have_selector("input.custom_class[name='q[created_at_lteq_datetime]']")
+        expect(body).to have_selector("input.custom_class[name='q[created_at_gteq]']")
+        expect(body).to have_selector("input.custom_class[name='q[created_at_lteq]']")
       end
     end
   end
@@ -252,38 +267,38 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
       let(:body) { filter :id }
 
       it "should generate a select option for equal to" do
-        expect(body).to have_selector("option[value=id_equals]", text: "Equals")
+        expect(body).to have_selector("option[value=id_eq]", text: "Equals")
       end
 
       it "should generate a select option for greater than" do
-        expect(body).to have_selector("option[value=id_greater_than]", text: "Greater than")
+        expect(body).to have_selector("option[value=id_gt]", text: "Greater than")
       end
 
       it "should generate a select option for less than" do
-        expect(body).to have_selector("option[value=id_less_than]", text: "Less than")
+        expect(body).to have_selector("option[value=id_lt]", text: "Less than")
       end
 
       it "should generate a text field for input" do
-        expect(body).to have_selector("input[name='q[id_equals]']")
+        expect(body).to have_selector("input[name='q[id_eq]']")
       end
 
       it "should select the option which is currently being filtered" do
-        scope = Post.ransack id_greater_than: 1
+        scope = Post.ransack id_gt: 1
         body = Capybara.string(render_filter scope, id: {})
-        expect(body).to have_selector("option[value=id_greater_than][selected=selected]", text: "Greater than")
+        expect(body).to have_selector("option[value=id_gt][selected=selected]", text: "Greater than")
       end
     end
 
     context "with filters options" do
-      let(:body) { filter :id, filters: [:equals, :greater_than] }
+      let(:body) { filter :id, filters: [:eq, :gt] }
 
       it "should generate provided options for filter select" do
-        expect(body).to have_selector("option[value=id_equals]", text: "Equals")
-        expect(body).to have_selector("option[value=id_greater_than]", text: "Greater than")
+        expect(body).to have_selector("option[value=id_eq]", text: "Equals")
+        expect(body).to have_selector("option[value=id_gt]", text: "Greater than")
       end
 
       it "should not generate a select option for less than" do
-        expect(body).not_to have_selector("option[value=id_less_than]")
+        expect(body).not_to have_selector("option[value=id_lt]")
       end
     end
   end
@@ -341,8 +356,8 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
 
       it "should generate a numeric filter" do
         expect(body).to have_selector("label", text: "Author") # really this should be Author ID :/)
-        expect(body).to have_selector("option[value=author_id_less_than]")
-        expect(body).to have_selector("input#q_author_id[name='q[author_id_equals]']")
+        expect(body).to have_selector("option[value=author_id_lt]")
+        expect(body).to have_selector("input#q_author_id[name='q[author_id_eq]']")
       end
     end
 
@@ -429,7 +444,7 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
   end # belongs to
 
   describe "has_and_belongs_to_many" do
-    skip "add HABTM models so this can be tested"
+    # skip "add HABTM models so this can be tested"
   end
 
   describe "has_many :through" do
@@ -491,26 +506,26 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
       context "with #{verb.inspect} proc" do
         it "#{should} be displayed if true" do
           body = filter :body, verb => proc { true }
-          expect(body).send if_true, have_selector("input[name='q[body_contains]']")
+          expect(body).send if_true, have_selector("input[name='q[body_cont]']")
         end
 
         it "#{should} be displayed if false" do
           body = filter :body, verb => proc { false }
-          expect(body).send if_false, have_selector("input[name='q[body_contains]']")
+          expect(body).send if_false, have_selector("input[name='q[body_cont]']")
         end
 
         it "should still be hidden on the second render" do
           filters = { body: { verb => proc { verb == :unless } } }
           2.times do
             body = Capybara.string(render_filter scope, filters)
-            expect(body).not_to have_selector("input[name='q[body_contains]']")
+            expect(body).not_to have_selector("input[name='q[body_cont]']")
           end
         end
 
         it "should successfully keep rendering other filters after one is hidden" do
           filters = { body: { verb => proc { verb == :unless } }, author: {} }
           body = Capybara.string(render_filter scope, filters)
-          expect(body).not_to have_selector("input[name='q[body_contains]']")
+          expect(body).not_to have_selector("input[name='q[body_cont]']")
           expect(body).to have_selector("select[name='q[author_id_eq]']")
         end
       end
@@ -520,9 +535,9 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
   describe "custom search methods" do
     it "should use the default type of the ransacker" do
       body = filter :custom_searcher_numeric
-      expect(body).to have_selector("option[value=custom_searcher_numeric_equals]")
-      expect(body).to have_selector("option[value=custom_searcher_numeric_greater_than]")
-      expect(body).to have_selector("option[value=custom_searcher_numeric_less_than]")
+      expect(body).to have_selector("option[value=custom_searcher_numeric_eq]")
+      expect(body).to have_selector("option[value=custom_searcher_numeric_gt]")
+      expect(body).to have_selector("option[value=custom_searcher_numeric_lt]")
     end
 
     it "should work as select" do
@@ -532,19 +547,19 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
 
     it "should work as string" do
       body = filter :custom_title_searcher, as: :string
-      expect(body).to have_selector("option[value=custom_title_searcher_contains]")
-      expect(body).to have_selector("option[value=custom_title_searcher_starts_with]")
+      expect(body).to have_selector("option[value=custom_title_searcher_cont]")
+      expect(body).to have_selector("option[value=custom_title_searcher_start]")
     end
 
     describe "custom date range search" do
       let(:gteq) { "2010-10-01" }
       let(:lteq) { "2010-10-02" }
-      let(:scope) { Post.ransack custom_created_at_searcher_gteq_datetime: gteq, custom_created_at_searcher_lteq_datetime: lteq }
+      let(:scope) { Post.ransack custom_created_at_searcher_gteq: gteq, custom_created_at_searcher_lteq: lteq }
       let(:body) { filter :custom_created_at_searcher, as: :date_range }
 
       it "should work as date_range" do
-        expect(body).to have_selector("input[name='q[custom_created_at_searcher_gteq_datetime]'][value='2010-10-01']")
-        expect(body).to have_selector("input[name='q[custom_created_at_searcher_lteq_datetime]'][value='2010-10-02']")
+        expect(body).to have_selector("input[name='q[custom_created_at_searcher_gteq]'][value='2010-10-01']")
+        expect(body).to have_selector("input[name='q[custom_created_at_searcher_lteq]'][value='2010-10-02']")
       end
 
       context "filter value can't be casted to date" do
@@ -552,8 +567,8 @@ RSpec.describe ActiveAdmin::Filters::ViewHelper do
         let(:lteq) { "Ooops" }
 
         it "should work display empty filter values" do
-          expect(body).to have_selector("input[name='q[custom_created_at_searcher_gteq_datetime]'][value='']")
-          expect(body).to have_selector("input[name='q[custom_created_at_searcher_lteq_datetime]'][value='']")
+          expect(body).to have_selector("input[name='q[custom_created_at_searcher_gteq]'][value='']")
+          expect(body).to have_selector("input[name='q[custom_created_at_searcher_lteq]'][value='']")
         end
       end
     end
