@@ -68,7 +68,7 @@ RSpec.describe ActiveAdmin::Filters::ResourceExtension do
 
     it "should raise an exception when filters are disabled" do
       resource.filters = false
-      expect { resource.remove_filter :author }.to raise_error ActiveAdmin::Filters::Disabled
+      expect { resource.remove_filter :author }.to raise_error(ActiveAdmin::Filters::Disabled, /Cannot remove a filter/)
     end
   end
 
@@ -116,7 +116,7 @@ RSpec.describe ActiveAdmin::Filters::ResourceExtension do
 
     it "should raise an exception when filters are disabled" do
       resource.filters = false
-      expect { resource.add_filter :title }.to raise_error ActiveAdmin::Filters::Disabled
+      expect { resource.add_filter :title }.to raise_error(ActiveAdmin::Filters::Disabled, /Cannot add a filter/)
     end
   end
 
