@@ -44,7 +44,7 @@ RSpec.describe ActiveAdmin::FormBuilder do
   end
 
   context "in general" do
-    context "it without custom settings" do
+    context "without custom settings" do
       let :body do
         build_form do |f|
           f.inputs do
@@ -59,10 +59,10 @@ RSpec.describe ActiveAdmin::FormBuilder do
       end
     end
 
-    context "it with custom settings" do
+    context "with custom settings" do
       let :body do
         build_form do |f|
-          f.inputs class: "custom_class", name: "custom_name", custom_attr: "custom_attr" do
+          f.inputs class: "custom_class", name: "custom_name", custom_attr: "custom_attr", data: { test: "custom" } do
             f.input :title
             f.input :body
           end
@@ -79,6 +79,29 @@ RSpec.describe ActiveAdmin::FormBuilder do
 
       it "should generate a fieldset with a custom attributes" do
         expect(body).to have_selector("fieldset[custom_attr='custom_attr']")
+      end
+
+      it "should use the rails helper for rendering attributes" do
+        expect(body).to have_css("fieldset[data-test='custom']")
+      end
+
+      it "should generate a legend with an inner span containing the name" do
+        expect(body).to have_selector("fieldset legend span", text: "custom_name")
+      end
+    end
+
+    context "with XSS payload as name" do
+      let :body do
+        build_form do |f|
+          f.inputs name: "<script>alert(document.domain)</script>" do
+            f.input :title
+            f.input :body
+          end
+        end
+      end
+
+      it "should generate a fieldset with the proper legend" do
+        expect(body).to have_css("legend", text: "<script>alert(document.domain)</script>")
       end
     end
   end
