@@ -54,7 +54,7 @@ RSpec.describe ActiveAdmin::Namespace do
     let(:namespace) { ActiveAdmin::Namespace.new(application, :admin) }
 
     it "should inherit the site title from the application" do
-      ActiveSupport::Deprecation.silence do
+      ActiveAdmin.deprecator.silence do
         ActiveAdmin::Namespace.setting :site_title, "Not the Same"
       end
       expect(namespace.site_title).to eq application.site_title
