@@ -109,9 +109,9 @@ RSpec.describe ActiveAdmin::Resource::Routes do
       end
 
       it "should include :scope and :q params" do
-        params = ActionController::Parameters.new(category_id: 1, q: { name_equals: "Any" }, scope: :all)
+        params = ActionController::Parameters.new(category_id: 1, q: { name_eq: "Any" }, scope: :all)
         additional_params = { locale: "en" }
-        batch_action_path = "/admin/categories/1/posts/batch_action?locale=en&q%5Bname_equals%5D=Any&scope=all"
+        batch_action_path = "/admin/categories/1/posts/batch_action?locale=en&q%5Bname_eq%5D=Any&scope=all"
 
         expect(config.route_batch_action_path(params, additional_params)).to eq batch_action_path
       end
@@ -122,9 +122,9 @@ RSpec.describe ActiveAdmin::Resource::Routes do
       let(:config) { ActiveAdmin.register News }
 
       it "should return the plural batch action route with _index and given params" do
-        params = ActionController::Parameters.new(q: { name_equals: "Any" }, scope: :all)
+        params = ActionController::Parameters.new(q: { name_eq: "Any" }, scope: :all)
         additional_params = { locale: "en" }
-        batch_action_path = "/admin/news/batch_action?locale=en&q%5Bname_equals%5D=Any&scope=all"
+        batch_action_path = "/admin/news/batch_action?locale=en&q%5Bname_eq%5D=Any&scope=all"
         expect(config.route_batch_action_path(params, additional_params)).to eq batch_action_path
       end
     end
