@@ -228,16 +228,43 @@ RSpec.describe ActiveAdmin::Views::PaginatedCollection do
       end
     end
 
-    it "makes no expensive COUNT queries when pagination_total is false" do
-      undecorated_collection = Post.all.page(1).per(30)
+    describe "when pagination_total is false" do
+      it "makes no expensive COUNT queries" do
+        undecorated_collection = Post.all.page(1).per(30)
 
-      expect { paginated_collection(undecorated_collection, pagination_total: false) }
-        .not_to perform_database_query("SELECT COUNT(*) FROM \"posts\"")
+        expect { paginated_collection(undecorated_collection, pagination_total: false) }
+          .not_to perform_database_query("SELECT COUNT(*) FROM \"posts\"")
 
-      decorated_collection = controller_with_decorator("index", PostDecorator).apply_collection_decorator(undecorated_collection)
+        decorated_collection = controller_with_decorator("index", PostDecorator).apply_collection_decorator(undecorated_collection.reset)
 
-      expect { paginated_collection(decorated_collection, pagination_total: false) }
-        .not_to perform_database_query("SELECT COUNT(*) FROM \"posts\"")
+        expect { paginated_collection(decorated_collection, pagination_total: false) }
+          .not_to perform_database_query("SELECT COUNT(*) FROM \"posts\"")
+      end
+
+      it "makes a performant COUNT query to figure out if we are on the last page" do
+        # "SELECT COUNT(*) FROM (SELECT 1". Let's make sure the subquery has LIMIT and OFFSET. It shouldn't have ORDER BY
+        count_query = %r{SELECT COUNT\(\*\) FROM \(SELECT 1 .*FROM "posts" (?=.*OFFSET \?)(?=.*LIMIT \?)(?!.*ORDER BY)}
+
+        undecorated_collection = Post.all.page(1).per(30)
+
+        expect { paginated_collection(undecorated_collection, pagination_total: false) }
+          .to perform_database_query(count_query)
+
+        undecorated_sorted_collection = undecorated_collection.reset.order(id: :desc)
+
+        expect { paginated_collection(undecorated_sorted_collection, pagination_total: false) }
+          .to perform_database_query(count_query)
+
+        decorated_collection = controller_with_decorator("index", PostDecorator).apply_collection_decorator(undecorated_collection.reset)
+
+        expect { paginated_collection(decorated_collection, pagination_total: false) }
+          .to perform_database_query(count_query)
+
+        decorated_sorted_collection = controller_with_decorator("index", PostDecorator).apply_collection_decorator(undecorated_sorted_collection.reset)
+
+        expect { paginated_collection(decorated_sorted_collection, pagination_total: false) }
+          .to perform_database_query(count_query)
+      end
     end
 
     it "makes no COUNT queries to figure out the last element of each page" do
